feat: replace UiRenderer with HeadlessSink, delete legacy modules
New:
- headless_sink.rs (146 lines) — simple println sink, auto-approves
- tool_history.rs (76 lines) — extracted ToolOutputHistory/Record

Deleted:
- confirm.rs (104 lines) — no longer referenced
- display.rs (922 lines) — print_ functions, tool_info, format_ helpers
- markdown.rs (564 lines) — streaming markdown renderer

Simplified:
- sink.rs: 340 → 37 lines (removed SinkMode::Direct, UiRenderer)
- headless.rs: uses HeadlessSink instead of CliSink::new()
- tui_commands.rs: inlined /expand output (was print_tool_output_full)
- repl.rs: inlined format_agents_list

Net: +265 -1,916 = -1,651 lines

Closes #85 phases 1, 3, 4, 5
